[General] restructuring folders
