+++ src/BlockStorageActions/RequestHelper.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace wappr\DigitalOcean\BlockStorageActions;
+
+trait RequestHelper
+{
+    // 
+}
