+++ src/Contracts/Droplets/DeleteDropletInterface.php
