+++ src/Factory.php
@@ -7,14 +7,14 @@ use ReflectionClass;
 use wappr\DigitalOcean\Contracts\FactoryInterface;
 
 /**
- * Class doFactory.
+ * Class Factory.
  *
  * @method static mixed create(string $actionType, array $params = [], Client $client = null)
  * @method static mixed retrieve(string $actionType, array $params = [], Client $client = null)
  * @method static mixed update(string $actionType, array $params = [], Client $client = null)
  * @method static mixed delete(string $actionType, array $params = [], Client $client = null)
  */
-class doFactory implements FactoryInterface
+class Factory implements FactoryInterface
 {
     /**
      * @param $name
